Styles & List Ordering
Added the ability to nest lists properly, adding LI wrappers to rogue list items, and melding style tags together with extremely deprecated attributes (like "bgcolor"). Added some fixups for the HEAD content area as well as the cutting for the BODY tags, in case for some reason they are upper-case. README update. This project is a mess.